requirements.txt
cairosvg
pillow
keyboard
evdev
torch
torchvision
opencv-python
